move task to shared module
