+++ src/app/shared-module/task-list/task-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { Task } from '../../home-page/types/task';
+import { Task } from '../types/task';
 
 @Component({
   selector: 'app-task-list',
