+++ src/app/statistics-page/components/statistics-page/statistics-page.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { TaskService } from 'src/app/shared-module/services/task.service';
-import { Task } from 'src/app/home-page/types/task';
+import { Task } from 'src/app/shared-module/types/task';
 
 @Component({
   selector: 'app-statistics-page',
